Change "localhost" to host ip addr
When the JavaScript code uses localhost, it's interpreted in the context of the user's browser, not the Docker containers. So if the user's browser is running on a different machine than the Docker host (even if that machine is a virtual machine on the same physical computer), localhost will not point to the Docker host.

So localhost needs to be replaced with the IP address of the Docker host in theJavaScript fetch call. The fetch call is made by the browser, not the Nginx or Flask containers, so it's the browser's network context that matters.